+++ src/main/java/ortus/boxlang/runtime/scopes/Key.java
@@ -211,6 +211,7 @@ public class Key implements Comparable<Key>, Serializable {
 	public static final Key		clientCertPassword					= Key.of( "clientCertPassword" );
 	public static final Key		cluster								= Key.of( "cluster" );
 	public static final Key		closure								= Key.of( "closure" );
+	public static final Key		codeKeys							= Key.of( "codeKeys" );
 	public static final Key		codePrintHTML						= Key.of( "codePrintHTML" );
 	public static final Key		codePrintPlain						= Key.of( "codePrintPlain" );
 	public static final Key		coldfusion							= Key.of( "coldfusion" );
@@ -363,6 +364,7 @@ public class Key implements Comparable<Key>, Serializable {
 	public static final Key		end									= Key.of( "end" );
 	public static final Key		endDate								= Key.of( "endDate" );
 	public static final Key		endTime								= Key.of( "endTime" );
+	public static final Key		enforceEncryptedSource				= Key.of( "enforceEncryptedSource" );
 	public static final Key		endRow								= Key.of( "endRow" );
 	public static final Key		endsWithNoCase						= Key.of( "endsWithNoCase" );
 	public static final Key		enforceExplicitOutput				= Key.of( "enforceExplicitOutput" );